analyzer/requirements.txt
requests
bs4
scapy-python3
matplotlib
numpy
elasticsearch
pyyaml
GeoIP
